+++ tox.ini
@@ -8,4 +8,3 @@ deps =
 
 commands =
     python -m unittest discover -s tests
-
